move MicroserviceRegistration class to internal.deployment.msf4j package
