+++ components/org.wso2.carbon.uis/src/main/java/org/wso2/carbon/uis/internal/deployment/listener/AppTransportBinder.java
@@ -23,9 +23,9 @@ import org.slf4j.LoggerFactory;
 import org.wso2.carbon.uis.api.App;
 import org.wso2.carbon.uis.api.ServerConfiguration;
 import org.wso2.carbon.uis.internal.deployment.AppDeploymentEventListener;
+import org.wso2.carbon.uis.internal.deployment.msf4j.MicroserviceRegistration;
 import org.wso2.carbon.uis.internal.exception.AppDeploymentEventListenerException;
 import org.wso2.carbon.uis.internal.http.RequestDispatcher;
-import org.wso2.carbon.uis.internal.http.msf4j.MicroserviceRegistration;
 import org.wso2.carbon.uis.internal.http.msf4j.MicroservicesRegistrar;
 import org.wso2.carbon.uis.internal.http.msf4j.WebappMicroservice;
