+++ components/org.wso2.carbon.uis/src/test/java/org/wso2/carbon/uis/internal/deployment/listener/AppTransportBinderTest.java
@@ -24,9 +24,9 @@ import org.testng.annotations.Test;
 import org.wso2.carbon.uis.api.App;
 import org.wso2.carbon.uis.api.Configuration;
 import org.wso2.carbon.uis.api.ServerConfiguration;
+import org.wso2.carbon.uis.internal.deployment.msf4j.MicroserviceRegistration;
 import org.wso2.carbon.uis.internal.exception.AppDeploymentEventListenerException;
 import org.wso2.carbon.uis.internal.http.HttpTransport;
-import org.wso2.carbon.uis.internal.http.msf4j.MicroserviceRegistration;
 import org.wso2.carbon.uis.internal.http.msf4j.MicroservicesRegistrar;
 
 import java.util.Collections;
