+++ components/org.wso2.carbon.uis/src/test/java/org/wso2/carbon/uis/internal/deployment/listener/RestApiDeployerTest.java
@@ -23,9 +23,9 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 import org.wso2.carbon.uis.api.App;
 import org.wso2.carbon.uis.api.ServerConfiguration;
+import org.wso2.carbon.uis.internal.deployment.msf4j.MicroserviceRegistration;
 import org.wso2.carbon.uis.internal.exception.AppDeploymentEventListenerException;
 import org.wso2.carbon.uis.internal.http.HttpTransport;
-import org.wso2.carbon.uis.internal.http.msf4j.MicroserviceRegistration;
 import org.wso2.carbon.uis.internal.http.msf4j.MicroservicesRegistrar;
 import org.wso2.carbon.uis.spi.RestApiProvider;
 import org.wso2.msf4j.Microservice;
